+++ r2l-api/src/builders/mod.rs
@@ -1,10 +1,7 @@
 pub mod agents;
 pub mod distribution;
 pub mod env;
-// pub mod env_pool;
 pub mod learning_module;
-// pub mod on_policy_algo;
 pub mod on_policy_algo2;
 pub mod sampler;
-// pub mod sampler_hooks;
 pub mod sampler_hooks2;
