+++ r2l-api/src/utils/evaluator.rs
@@ -1,15 +1,12 @@
 use candle_core::{Device, Result, Tensor};
 use r2l_core::{
-    distributions::Distribution,
-    env::Env,
-    // env_pools::{SequentialVecEnvHooks, single_step_env_with_buffer},
-    numeric::Buffer,
-    utils::rollout_buffer::RolloutBuffer,
+    distributions::Distribution, env::Env, numeric::Buffer,
+    sampler::trajectory_buffers::variable_size_buffer::VariableSizedTrajectoryBuffer,
 };
 use std::sync::{Arc, Mutex};
 
 pub struct Evaluator<E: Env> {
-    pub env: E,
+    pub trajectory_buffer: VariableSizedTrajectoryBuffer<E>,
     pub eval_episodes: usize,
     pub eval_freq: usize,
     pub eval_step: usize,
@@ -26,7 +23,7 @@ impl<E: Env<Tensor = Buffer>> Evaluator<E> {
         evaluations_results: Arc<Mutex<Vec<Vec<f32>>>>,
     ) -> Self {
         Self {
-            env,
+            trajectory_buffer: VariableSizedTrajectoryBuffer::new(env),
             eval_episodes,
             eval_freq,
             eval_step,
@@ -49,55 +46,27 @@ impl<E: Env<Tensor = Buffer>> Evaluator<E> {
             Ok(())
         } else {
             // TODO: VariableSizedTrajectoryBuf
-            //
-            // let mut all_rewards = vec![];
-            // let mut dones = vec![];
-            // for _ in 0..self.eval_episodes {
-            //     let mut state = self
-            //         .env
-            //         .reset(rand::random())
-            //         .to_candle_tensor(&Device::Cpu);
-            //     let mut rollout_buffer = RolloutBuffer::default();
-            //     while let (next_state, false) = single_step_env_with_buffer(
-            //         dist,
-            //         &state,
-            //         &mut self.env,
-            //         &mut rollout_buffer,
-            //         &self.device,
-            //     )? {
-            //         state = next_state;
-            //     }
-            //     all_rewards.push(rollout_buffer.rewards.clone());
-            //     dones.push(rollout_buffer.dones.clone());
-            // }
-            // let eps = dones
-            //     .iter()
-            //     .map(|x| x.iter().filter(|y| **y).count())
-            //     .sum::<usize>();
-            // let sum_rewards_per_eps: Vec<f32> =
-            //     all_rewards.iter().map(|x| x.iter().sum::<f32>()).collect();
-            // let avg_rewards = sum_rewards_per_eps.iter().sum::<f32>() / eps as f32;
-            // println!("Avg rew: {}", avg_rewards);
-            // let mut eval_results = self.evaluations_results.lock().unwrap();
-            // eval_results.push(sum_rewards_per_eps);
-            // self.eval_step = 0;
+            let mut all_rewards = vec![];
+            let mut dones = vec![];
+            for _ in 0..self.eval_episodes {
+                // TODO: variable sized buffer should be able to handle single episodes
+                self.trajectory_buffer.run_episode(dist, 1);
+                let rollout_buffer = self.trajectory_buffer.to_rollout_buffer();
+                all_rewards.push(rollout_buffer.rewards.clone());
+                dones.push(rollout_buffer.dones.clone());
+            }
+            let eps = dones
+                .iter()
+                .map(|x| x.iter().filter(|y| **y).count())
+                .sum::<usize>();
+            let sum_rewards_per_eps: Vec<f32> =
+                all_rewards.iter().map(|x| x.iter().sum::<f32>()).collect();
+            let avg_rewards = sum_rewards_per_eps.iter().sum::<f32>() / eps as f32;
+            println!("Avg rew: {}", avg_rewards);
+            let mut eval_results = self.evaluations_results.lock().unwrap();
+            eval_results.push(sum_rewards_per_eps);
+            self.eval_step = 0;
             Ok(())
         }
     }
 }
-
-// impl<E: Env<Tensor = Buffer>> SequentialVecEnvHooks for Evaluator<E> {
-//     fn step_hook(
-//         &mut self,
-//         distribution: &dyn Distribution<Tensor = Tensor>,
-//         states: &mut Vec<(Tensor, Tensor, f32, bool)>,
-//     ) -> candle_core::Result<bool> {
-//         let n_envs = states.len();
-//         self.evaluate(distribution, n_envs)?;
-//         Ok(false)
-//     }
-//
-//     fn post_step_hook(&mut self, last_states: &mut Vec<Tensor>) -> candle_core::Result<bool> {
-//         Ok(false)
-//     }
-// }
