+++ r2l-core/src/lib.rs
@@ -27,7 +27,6 @@
 pub mod agents;
 pub mod distributions;
 pub mod env;
-// pub mod env_pools;
 pub mod ipc;
 pub mod numeric;
 pub mod on_policy_algorithm;
